requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.2.1
Django==3.1.6
django-extensions==3.1.1
django-shell-plus==1.1.7
djangorestframework==3.12.2
jwt==1.1.0
numpy==1.19.5
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
